refactor: inline ZeroTalkPage wrapper into ZeroChatPage
Merge the thin ZeroTalkPage wrapper into ZeroChatPage so the talk page
follows the same single-component pattern as other pages. ZeroChatPage
now owns agent resolution, send logic, and SidebarLayout wrapping
directly, eliminating the unnecessary indirection layer.

Closes #7411
